Check path instead of catching error

+++ rapthor/modifystate.py
@@ -128,15 +128,12 @@ def run(parset_file):
                 "images",
                 "visibilities",
             ):
-                dcmp = filecmp.dircmp(
-                    os.path.join(parset['dir_working'], 'pipelines'),
-                    os.path.join(parset['dir_working'], dirname)
-                )
-                try:
+                if os.path.exists(os.path.join(parset['dir_working'], dirname)):
+                    dcmp = filecmp.dircmp(
+                        os.path.join(parset['dir_working'], 'pipelines'),
+                        os.path.join(parset['dir_working'], dirname)
+                    )
                     for path in (os.path.join(dcmp.right, item) for item in dcmp.right_only):
                         shutil.rmtree(path, ignore_errors=True)
-                except FileNotFoundError:
-                    # Directory does not exist
-                    continue
 
             print('Reset complete.')
